Modified method "removeById" in PostRepository.java and modified "GET" and "DELETE" in MainServlet.java

+++ src/main/java/ru/netology/repository/PostRepository.java
@@ -46,7 +46,10 @@ public class PostRepository {
     }
 
     public void removeById(long id) {
-        repository.remove(id);
-        throw new NotFoundException();
+        if(repository.containsKey(id)) {
+            repository.remove(id);
+        } else {
+            throw new NotFoundException();
+        }
     }
 }
